Refactor padding in Delivery_schedule.css and fix login page layout
The padding in the .DScontainer class in Delivery_schedule.css was adjusted to provide more space. Additionally, the conflict in the Login.js file was resolved, removing unnecessary code related to the "Forgot password" feature.

Refactor SQL triggers and procedures for order management

The SQLtriggers.sql file was modified to add a trigger that checks for available space before updating the Train table. The ManagerModel.js file was also updated to improve the addTrainDelivery function and add console logs for debugging purposes.

Update AssistantWorkedHours.js to display correct hours worked

The AssistantWorkedHours.js file was modified to display the correct hours worked by assistants in the table.

Update SQLprocedures.sql to add createTrainSchedule procedure

The SQLprocedures.sql file was updated to add the createTrainSchedule procedure, which inserts a new record into the train_delivery table and updates the Vehicle_departure_time in the Delivery_Schedule table.

Refactor package-lock.json and remove license information

The package-lock.json file was refactored and the license information was removed.

Merge pull request #112 from hirunanimesh/heshan3

This is a merge commit from the pull request #112.

Merge pull request #111 from hirunanimesh/hiru31

This is a merge commit from the pull request #111.